Moved policy struct to new file and created new function to get current password policy
